pyproject.toml
[project]
authors = [
    {name = "Dmitry Kolomatskiy", email = "58207913+lemontree210@users.noreply.github.com"},
]
requires-python = "<3.11, >=3.10"
dependencies = [
    "django<6.0,>=5.0",
    "gunicorn>=22.0.0,<24.0.0",
    "django-phonenumber-field[phonenumberslite]<9.0.0,>=8.0.0",
    "django-extensions>=3.2.3,<5.0.0",
    "psycopg2-binary<3.0.0,>=2.9.9",
    "djangorestframework==3.16.0",
    "drf-spectacular<1.0.0,>=0.27.2",
    "django-filter>=24.2,<26.0",
    "django-stubs-ext==5.1.3",
    "django-reversion<6.0.0,>=5.0.12",
    "email-validator<3.0.0,>=2.2.0",
    "geopy<3.0.0,>=2.4.1",
    "timezonefinder<7.0.0,>=6.5.2",
    "pytz>=2024.1,<2026.0",
    "tqdm<5.0.0,>=4.66.4",
    "phonenumbers>=8.13.42,<10.0.0",
    "celery>=5.5.0",
    "redis>=5.2.1",
    "django-celery-beat>=2.7.0",
    "django-select2>=8.4.0",
    "django-redis>=5.4.0",
    "python-dotenv>=1.1.0",
    "pytest-mock>=3.14.1",
]
name = "django-webapps"
version = "0.1.0"
description = ""

[dependency-groups]
dev = [
    "black==24.10.0",
    "ruff==0.11.0",
    "pre-commit<4.0.0,>=3.8.0",
    "icecream<3.0.0,>=2.1.3",
    "isort<6.0.0,>=5.13.2",
    "graphviz<1.0.0,>=0.20.3",
    "mypy==1.15.0",
    "django-filter-stubs<1.0.0,>=0.1.3",
    "djangorestframework-stubs==3.15.0",
    "django-stubs==5.0.2",
    "pytest<9.0.0,>=8.2.2",
    "pytest-django<5.0.0,>=4.8.0",
    "pytest-asyncio<1.0.0,>=0.23.7",
    "faker<27.0.0,>=26.0.0",
    "model-bakery<2.0.0,>=1.18.2",
]

[tool.black]
line-length = 120
target-version = ['py310']
exclude = ".venv|api/migrations|docs"

[tool.django-stubs]
django_settings_module = "django_webapps.settings"

[tool.isort]
profile = "black"
line_length = 99
skip_glob = ["api/migrations/*"]

[tool.mypy]
exclude = ".venv|api/migrations|docs|devtools|tests|api/processors/actions/group/finish.py"
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
python_version = "3.10"
strict = true
disallow_untyped_calls = false
warn_return_any = false
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = [
  "tests.*"
]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
  "api.views.*"
]
follow_imports = "skip"
ignore_errors = true



[mypy.plugins.django-stubs]
django_settings_module = "api.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_webapps.settings"
asyncio_mode = "auto"
# short output to checke github actions output
addopts = "-ra -v --tb=short"

[tool.ruff]
line-length = 120
exclude = ["api/migrations"]
lint.per-file-ignores = {"__init__.py" = ["F401"]}
# Pyflakes and pycodestyle Error (included by default without explicit "select"),
# I = isort, N = pep8-naming, PL = pylint, UP = pyupgrade
# A = flake8-builtins, DJ = flake8-django C4 = flake8-comprehensions, DTZ = flake8-datetimez,
# ISC = flake8-implicit-str-concat, T20 = flake8-print, RET = flake8-return, SIM = flake8-simplify,
# ARG = flake8-unused-arguments, PTH = flake8-use-pathlib
lint.select = ["E", "F", "I", "N", "PL", "UP", "A", "DJ", "C4", "DTZ", "ISC", "T20", "RET", "SIM", "ARG", "PTH"]
# can add ERA (eradicate) later in here ^ for deleting commented out code
# Ignore: A003 (class attribute `id`/`type` is shadowing a python builtin) because class attributes don't really shadow Python builtins
lint.ignore = ["A003"]

# Assume Python 3.10.
target-version = "py310"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
